Fix a bug which could lead to the entire panel freezing
BEFORE:
When the applet was downloading the album cover image,
the download ran on the same thread as the rest of the applet
and the gtk loop. So on slow internet connection the entire panel
could freeze for multiple seconds while the image was downloading.

NOW:
Now, a new thread (using python's threading module) is created,
the download then happens in the new thread, not slowing
the main thread.