Moved settings handling for PanelControlView.py to a better place
Previously settings handling for the panel view (PanelControlView) was done in BudgieMediaPlayer
and then sent to the panelControlView through a set of methods, but this was complicated and unreadable.

So I moved the settings handling to PanelControlview which made the code more readable, shorter and less spaghetti.

Also the ellipting of the labels is now done trough gtk indead of being done manually,
which makes the code simpler and was required to do to make the canges above.