Fixed a bug where clicks were sometimes not registered
If the user clicked in the empty space between the two elements to would do nothing,
because this space would be part of the parent container.

This is fixed by setting margins on the elements instead of using the container spacing.
Unfortunately gtk doesnt register clicks on margins either, so the margins have to be set
on a child of an container that accepts clicks (like EventBox)